+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Version NEXT
 * Made Backends internationalizable, as well as the BillingShippingForm
   thanks to the introduciton of a new optional backend_verbose_name attribute
   to backends.
-* Added a order_required decorator to fix bug #84
+* Added a order_required decorator to fix bug #84, which should be used on all payment and shipping views
 * Added get_product_reference method to Product (for extensibility)
 * Cart object is not saved to database if it is empty (#147)
 * Changed spelling mistakes in methods from `payed` to `paid` on the Order